[BUGFIX] Assure clean environment when testing in CI
